Commented out some lines in Main that caused errors since they haven't been updated to work with the split XML file (still have the data one).
Some comments added and a couple of things regarding navigation and the
games. There's an issue that the music doesn't stop if you press the
DEVICE's home button from within a game. Fix is known, but then the
music will start from the beginning when entering the next game
(quiz-tap or vice versa).